refactor: modularize error module into separate components
Split the monolithic error.rs (688 lines) into focused modules:
- types.rs: EllError enum and core trait implementations
- builders.rs: Helper functions and error constructors
- sourceloc.rs: SourceLoc struct for tracking source positions
- runtime.rs: RuntimeError struct with location/context support
- mod.rs: Public API re-exports and full test suite (45 tests)

Benefits:
- Separation of concerns: type definitions, builders, and utilities isolated
- SourceLoc is now in its own focused module
- RuntimeError logic is independent and testable
- Cleaner module organization improves discoverability
- All error handling logic remains accessible via public API

Testing:
✅ All 1395 tests pass
✅ All 45 error-specific tests pass
✅ No functional changes
✅ Clippy: clean
✅ Format: verified
